Episode4/requirements.txt
streamlit
pandas
plotly
altair
Pillow
numpy
